server/requirements.txt
Django==6.0.3
djangorestframework==3.17.0
django-cors-headers==4.9.0
djangorestframework_simplejwt==5.5.1
channels==4.1.0
daphne==4.1.2
channels-redis==4.2.0
whitenoise==6.9.0
python-decouple==3.8
uvicorn==0.34.2
redis==6.0.0
psycopg2-binary==2.9.10
django-cloudinary-storage==0.3.0
Pillow==12.2.0